feat: Move controllers into separate modules
