+++ package.json
@@ -21,6 +21,7 @@
     "@typescript-eslint/parser": "4.28.3",
     "eslint": "7.31.0",
     "eslint-plugin-import": "2.23.4",
+    "prettier": "^2.3.2",
     "ts-node": "10.1.0",
     "typescript": "4.3.5"
   }
