+++ src/core/common/schemas.ts
